React hook to handle manual refresh state
