+++ frontend/src/pages/Overview/OverviewPage.tsx
@@ -2,12 +2,10 @@ import * as React from 'react';
 import { Grid, GridItem } from '@patternfly/react-core';
 import { kialiStyle } from 'styles/StyleUtils';
 import { serverConfig } from 'config/ServerConfig';
-import { RefreshIntervalManual } from 'config/Config';
 import { DefaultSecondaryMasthead } from 'components/DefaultSecondaryMasthead/DefaultSecondaryMasthead';
 import { Refresh } from 'components/Refresh/Refresh';
 import { useKialiDispatch } from 'hooks/redux';
-import { useRefreshInterval } from 'hooks/refresh';
-import { HistoryManager } from 'app/History';
+import { useManualRefreshState } from 'hooks/refresh';
 import { ManualRefreshEmptyState } from 'components/Refresh/ManualRefreshEmptyState';
 import { setAIContext } from 'helpers/ChatAI';
 import { t } from 'utils/I18nUtils';
@@ -48,37 +46,11 @@ const secondRowItemStyle = kialiStyle({
 
 export const OverviewPage: React.FC = () => {
   const dispatch = useKialiDispatch();
-  const { lastRefreshAt, refreshInterval } = useRefreshInterval();
-
-  // Check multiple sources for manual mode because no single source is reliable in all cases:
-  // - Redux: correct for client-side navigation (persists across pages) but stale on F5 (default value)
-  // - URL (HistoryManager/window.location): correct on F5 but empty on client-side navigation
-  //   (search params don't carry across React Router navigations)
-  const [loaded, setLoaded] = React.useState(
-    () => refreshInterval !== RefreshIntervalManual && HistoryManager.getRefresh() !== RefreshIntervalManual
-  );
-
-  const initialRefreshAt = React.useRef(lastRefreshAt);
-  const prevRefreshInterval = React.useRef(refreshInterval);
+  const loaded = useManualRefreshState();
 
   React.useEffect(() => {
     setAIContext(dispatch, 'Overview page');
   }, [dispatch]);
-
-  React.useEffect(() => {
-    if (!loaded && lastRefreshAt !== initialRefreshAt.current) {
-      setLoaded(true);
-    }
-  }, [lastRefreshAt, loaded]);
-
-  React.useEffect(() => {
-    const prev = prevRefreshInterval.current;
-    prevRefreshInterval.current = refreshInterval;
-
-    if (!loaded && prev === RefreshIntervalManual && refreshInterval !== RefreshIntervalManual) {
-      setLoaded(true);
-    }
-  }, [refreshInterval, loaded]);
 
   const durationLabel = serverConfig.healthConfig?.compute?.duration ?? '5m';
